Removing autoloading of app/helpers paths from configuration_extensions.
Including the app/helpers path was causing all helpers to be loaded at all times. In case of Issue #56, the result was that the meta_errors? method from pages_helper in all cases, even when we should have been using the base application_helper version of the method. Removing app/helpers from the load path took care of this, but I'm concerned this may cause other issues. Left a TODO in the code as a flag to look into this further.