Adding a render command to the show_page controller action to render actual pages.
Previous, some kind of Rails2 magic (or hack) was being invoked, and at the model level, request and response objects where being manipulated. Somehow, this was overriding Rails' default Fine Me A View That Matches This Route function. In Rails 3: not so much. (It's also possible the magic is a monkeypatch somewhere withing Radiant's codebase.) Either way, now we've created a layoutless view that simply renders the response.body, contianing the page's full HTML. It seems to be working. Closes Issue #74.