Add attr_accessibles for everything we have tests for
This article recommends we prefer the whitelist using attr_accessible
instead of attr_protected:
http://code.tutsplus.com/tutorials/mass-assignment-rails-and-you--net-31695

which sounds like a good idea even though we probably won't be adding
many new attributes.

It looks like radiant implemented their own version of this on User, so
I removed that and made it use the built in rails stuff instead. Good
for them!

There's probably more out there, but the tests aren't showing them to
me. At least the previous commit will make it obvious when we hit them.