Removed test to see if a column was already :text mid-migration.
The migrations add allowed_children_cache in one migration, then test to see if it's text before doing a change_column in another migration. For some reason, even if I tried to reset_column_information, that migration got nil trying to find the added column. I have no idea why.

I removed the check, and the migration works fine. So, we're going with that. Issue #98