Tests seem to be fixed to input text into CKEditor fields.
I'm not quite familiar enough with Capybara's execute_script and evaluate_script methods (and documentation on them is pretty thin) to know if I've done something Not The Way I Should, but this seems to be working now. @carols10cents, if you have a moment (since you know Capybara better than I do) could you take a look at these changes and see if I'm in the right vicinity of how to do this?

Issue #132