Beginning to set up the generator to copy all config files for a new project.
The old generators are all built for Rails 2 (including using the Rails 2 config file structure, which is very different in 3), so rather than port them, I'm rewriting from scratch.

This task will copy config files, but the command that generates the project itself still needs to be either ported or rewritten.
Issue #136