Moved FileNotFoundPage logic outside of conditional in find_by_path.
FileNotFoundPages were only searched in cases where the path being searched existed under the page in question.

In other words, if path was /page, but we were searching under a page with the slug home, we'd neither find /page nor would a FileNotFoundPage be searched for. On the other hand, if we searched for /home/page under /home, but /home/page didn't exist, a FileNotFoundPage would be searched for in its place and returned.

Now, that logic exists outside of that conditional. If a page is found in the existing conditionals, it's returned. If it isn't, we search for a FileNotFoundPage and return that instead.

This is a fairly significant change and needs to be tested thoroughly, but we're checking this in and deploying it so we can test it on a number of existing sites.