Making changes to old migrations. Yes, a dangerous idea, but read on...
The first migration creates the users table with integer values for what should be booleans. A later migration updates this. This breaks DB bootstrap for sites with extensions like multi-site that monkeypatch existing models when those extensions have yet to add their additional fields.

So, to avoid this happening (and considering the existing scope of highly breaking changes on the route to 2.0) I've updated the initial table creation for users and removed the migration that did the swap.