BugzID: 3529 - Implements jQuery Sortable
created a js file to sort the tables
I added some configs to ensure the placement of the tr doesn't go above the folder it's in and cannot be moved above the absolute parent folder. The top level folder cannot be moved. An expanded folder cannot be moved.

I also set up some bones in order to submit the position and save it. This will submit via ajax.